+++ packages/autocomplete/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.20](https://github.com/rentalutions/elements/compare/@rent_avail/autocomplete@0.3.16...@rent_avail/autocomplete@0.3.20) (2022-10-03)
+
+**Note:** Version bump only for package @rent_avail/autocomplete
+
+
+
+
+
+## [0.3.19](https://github.com/rentalutions/elements/compare/@rent_avail/autocomplete@0.3.18...@rent_avail/autocomplete@0.3.19) (2022-09-27)
+
+**Note:** Version bump only for package @rent_avail/autocomplete
+
+
+
+
+
+## [0.3.18](https://github.com/rentalutions/elements/compare/@rent_avail/autocomplete@0.3.16...@rent_avail/autocomplete@0.3.18) (2022-09-16)
+
+**Note:** Version bump only for package @rent_avail/autocomplete
+
+
+
+
+
+## [0.3.17](https://github.com/rentalutions/elements/compare/@rent_avail/autocomplete@0.3.16...@rent_avail/autocomplete@0.3.17) (2022-08-31)
+
+**Note:** Version bump only for package @rent_avail/autocomplete
+
+
+
+
+
+## [0.3.16](https://github.com/rentalutions/elements/compare/@rent_avail/autocomplete@0.3.15...@rent_avail/autocomplete@0.3.16) (2022-04-18)
+
+**Note:** Version bump only for package @rent_avail/autocomplete
+
+
+
+
+
 ## [0.3.15](https://github.com/rentalutions/elements/compare/@rent_avail/autocomplete@0.3.14...@rent_avail/autocomplete@0.3.15) (2021-11-08)
 
 **Note:** Version bump only for package @rent_avail/autocomplete